# desc/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Fishing Notes Admin API",
    "version": "0.2.1",
    "description": "管理后台 API 文档，覆盖认证、RBAC、管理员用户、钓点审核、知识专栏、鱼种内容与业务用户管理。\n\n使用说明：\n1. 除登录接口外，其余接口默认都需要 Bearer Token。\n2. 所有成功响应统一为 `{ code, message, data }`，其中 `code = 0` 表示成功。\n3. 列表接口统一支持 `page` 和 `pageSize`，默认 `page=1`、`pageSize=20`。\n"
  },
  "servers": [
    {
      "url": "http://localhost:1002",
      "description": "Local API Server"
    }
  ],
...
  "security": [
    {
      "bearerAuth": [

      ]
    }
  ],
  "paths": {
...
    "/api/v1/spot-corrections": {
      "get": {
        "tags": [
          "Spot Corrections"
        ],
        "summary": "获取钓点纠错列表",
        "operationId": "listSpotCorrections",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          },
          {
            "name": "keyword",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "approved",
                "rejected"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListSpotCorrectionsEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/v1/spot-corrections/{correctionId}": {
      "get": {
        "tags": [
          "Spot Corrections"
        ],
        "summary": "获取钓点纠错详情",
        "operationId": "getSpotCorrection",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CorrectionIdPath"
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpotCorrectionEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/v1/spot-corrections/{correctionId}/approve": {
      "post": {
        "tags": [
          "Spot Corrections"
        ],
        "summary": "采纳钓点纠错",
        "operationId": "approveSpotCorrection",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CorrectionIdPath"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewSpotCorrectionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/v1/spot-corrections/{correctionId}/reject": {
      "post": {
        "tags": [
          "Spot Corrections"
        ],
        "summary": "驳回钓点纠错",
        "operationId": "rejectSpotCorrection",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CorrectionIdPath"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewSpotCorrectionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/v1/media-assets": {
      "get": {
        "tags": ["Media Review"],
        "summary": "获取公开渔获图片审核队列",
        "operationId": "listMediaAssets",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"$ref": "#/components/parameters/Page"},
          {"$ref": "#/components/parameters/PageSize"},
          {"name": "keyword", "in": "query", "schema": {"type": "string"}},
          {"name": "status", "in": "query", "schema": {"type": "string", "enum": ["pending", "approved", "rejected"]}}
        ],
        "responses": {
          "200": {"description": "查询成功", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ListMediaAssetsEnvelope"}}}},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "500": {"$ref": "#/components/responses/InternalServerError"}
        }
      }
    },
    "/api/v1/media-assets/{mediaId}/approve": {
      "post": {
        "tags": ["Media Review"],
        "summary": "通过图片审核",
        "operationId": "approveMediaAsset",
        "security": [{"bearerAuth": []}],
        "parameters": [{"$ref": "#/components/parameters/MediaIdPath"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ReviewMediaAssetRequest"}}}},
        "responses": {
          "200": {"description": "操作成功", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SuccessEnvelope"}}}},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "500": {"$ref": "#/components/responses/InternalServerError"}
        }
      }
    },
    "/api/v1/media-assets/{mediaId}/reject": {
      "post": {
        "tags": ["Media Review"],
        "summary": "驳回图片审核",
        "operationId": "rejectMediaAsset",
        "security": [{"bearerAuth": []}],
        "parameters": [{"$ref": "#/components/parameters/MediaIdPath"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ReviewMediaAssetRequest"}}}},
        "responses": {
          "200": {"description": "操作成功", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SuccessEnvelope"}}}},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "500": {"$ref": "#/components/responses/InternalServerError"}
        }
      }
    },
    "/api/v1/public-spot-applications": {
      "get": {
        "tags": ["Public Spot Applications"],
        "summary": "查询公共钓点申请",
        "operationId": "listPublicSpotApplications",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"$ref": "#/components/parameters/Page"},
          {"$ref": "#/components/parameters/PageSize"},
          {"name": "keyword", "in": "query", "schema": {"type": "string"}},
          {"name": "status", "in": "query", "schema": {"type": "string", "enum": ["pending", "approved", "rejected"]}}
        ],
        "responses": {
          "200": {"description": "查询成功", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ListPublicSpotApplicationsEnvelope"}}}},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "500": {"$ref": "#/components/responses/InternalServerError"}
        }
      }
    },
    "/api/v1/public-spot-applications/{applicationId}": {
      "get": {
        "tags": ["Public Spot Applications"],
        "summary": "查询公共钓点申请详情",
        "operationId": "getPublicSpotApplication",
        "security": [{"bearerAuth": []}],
        "parameters": [{"$ref": "#/components/parameters/PublicSpotApplicationIdPath"}],
        "responses": {
          "200": {"description": "查询成功", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PublicSpotApplicationEnvelope"}}}},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "500": {"$ref": "#/components/responses/InternalServerError"}
        }
      }
    },
    "/api/v1/public-spot-applications/{applicationId}/approve": {
      "post": {
        "tags": ["Public Spot Applications"],
        "summary": "通过公共钓点申请",
        "operationId": "approvePublicSpotApplication",
        "security": [{"bearerAuth": []}],
        "parameters": [{"$ref": "#/components/parameters/PublicSpotApplicationIdPath"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ReviewPublicSpotApplicationRequest"}}}},
        "responses": {
          "200": {"description": "操作成功", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SuccessEnvelope"}}}},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "500": {"$ref": "#/components/responses/InternalServerError"}
        }
      }
    },
    "/api/v1/public-spot-applications/{applicationId}/reject": {
      "post": {
        "tags": ["Public Spot Applications"],
        "summary": "驳回公共钓点申请",
        "operationId": "rejectPublicSpotApplication",
        "security": [{"bearerAuth": []}],
        "parameters": [{"$ref": "#/components/parameters/PublicSpotApplicationIdPath"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ReviewPublicSpotApplicationRequest"}}}},
        "responses": {
          "200": {"description": "操作成功", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SuccessEnvelope"}}}},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "500": {"$ref": "#/components/responses/InternalServerError"}
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "Page": {
        "name": "page",
        "in": "query",
        "description": "页码，默认 1",
        "schema": {
          "type": "integer",
          "format": "int64",
          "default": 1
        }
      },
      "PageSize": {
        "name": "pageSize",
        "in": "query",
        "description": "每页数量，默认 20，最大 100",
        "schema": {
          "type": "integer",
          "format": "int64",
          "default": 20,
          "maximum": 100
        }
      },
...
      "CorrectionIdPath": {
        "name": "correctionId",
        "in": "path",
        "required": true,
        "description": "钓点纠错 ID",
        "schema": {
          "type": "integer",
          "format": "int64"
        }
      },
      "MediaIdPath": {
        "name": "mediaId",
        "in": "path",
        "required": true,
        "description": "图片资源 ID",
        "schema": {
          "type": "integer",
          "format": "int64"
        }
      },
      "PublicSpotApplicationIdPath": {
        "name": "applicationId",
        "in": "path",
        "required": true,
        "description": "公共钓点申请 ID",
        "schema": {"type": "integer", "format": "int64"}
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "请求参数错误",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "未认证或认证失败",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Forbidden": {
        "description": "没有权限访问",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "NotFound": {
        "description": "资源不存在",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "服务端异常",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      }
    },
    "schemas": {
...
      "SpotCorrectionItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "correctionCode": {
            "type": "string"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "userNickname": {
            "type": "string"
          },
          "spotCode": {
            "type": "string"
          },
          "spotName": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "approved",
              "rejected"
            ]
          },
          "reviewNote": {
            "type": "string"
          },
          "reviewerUserId": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "reviewedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "diff": {
            "$ref": "#/components/schemas/SpotCorrectionDiff"
          }
        }
      },
      "SpotCorrectionFieldChange": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "current": {
            "type": "string"
          },
          "proposed": {
            "type": "string"
          }
        }
      },
      "SpotCorrectionDiff": {
        "type": "object",
        "properties": {
          "structured": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "fields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpotCorrectionFieldChange"
            }
          }
        }
      },
      "ListSpotCorrectionsResponse": {
        "type": "object",
        "properties": {
          "list": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpotCorrectionItem"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "page": {
            "type": "integer",
            "format": "int64"
          },
          "pageSize": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "MediaAssetItem": {
        "type": "object",
        "properties": {
          "id": {"type": "integer", "format": "int64"},
          "recordCode": {"type": "string"},
          "userId": {"type": "integer", "format": "int64"},
          "spotName": {"type": "string"},
          "imageUrl": {"type": "string"},
          "width": {"type": "integer"},
          "height": {"type": "integer"},
          "moderationStatus": {"type": "string", "enum": ["pending", "approved", "rejected"]},
          "moderationNote": {"type": "string"},
          "createdAt": {"type": "string", "format": "date-time"}
        }
      },
      "ListMediaAssetsResponse": {
        "type": "object",
        "properties": {
          "list": {"type": "array", "items": {"$ref": "#/components/schemas/MediaAssetItem"}},
          "total": {"type": "integer", "format": "int64"},
          "page": {"type": "integer", "format": "int64"},
          "pageSize": {"type": "integer", "format": "int64"}
        }
      },
      "ReviewMediaAssetRequest": {
        "type": "object",
        "required": ["reviewNote"],
        "properties": {
          "reviewNote": {"type": "string", "maxLength": 500}
        }
      },
      "ReviewPublicSpotApplicationRequest": {
        "type": "object",
        "required": ["reviewNote"],
        "properties": {
          "reviewNote": {"type": "string", "maxLength": 500},
          "mergeSpotId": {"type": "integer", "format": "int64", "description": "通过时可选；提供后将申请并入该公共钓点，不提供则新建钓点"}
        }
      },
      "PublicSpotApplicationItem": {
        "type": "object",
        "properties": {
          "id": {"type": "integer", "format": "int64"},
          "applicationCode": {"type": "string"},
          "userId": {"type": "integer", "format": "int64"},
          "userNickname": {"type": "string"},
          "sourceSpotCode": {"type": "string"},
          "name": {"type": "string"},
          "address": {"type": "string"},
          "latitude": {"type": "number", "format": "double"},
          "longitude": {"type": "number", "format": "double"},
          "description": {"type": "string"},
          "chargeType": {"type": "string"},
          "targetSpecies": {"type": "array", "items": {"type": "string"}},
          "imageUrls": {"type": "array", "items": {"type": "string"}},
          "status": {"type": "string", "enum": ["pending", "approved", "rejected"]},
          "reviewNote": {"type": "string"},
          "mergedSpotId": {"type": "integer", "format": "int64"},
          "mergedSpotCode": {"type": "string"},
          "createdAt": {"type": "string", "format": "date-time"}
        }
      },
      "ListPublicSpotApplicationsResponse": {
        "type": "object",
        "properties": {
          "list": {"type": "array", "items": {"$ref": "#/components/schemas/PublicSpotApplicationItem"}},
          "total": {"type": "integer", "format": "int64"},
          "page": {"type": "integer", "format": "int64"},
          "pageSize": {"type": "integer", "format": "int64"}
        }
      },
...
      "ReviewSpotCorrectionRequest": {
        "type": "object",
        "required": [
          "reviewNote"
        ],
        "properties": {
          "reviewNote": {
            "type": "string",
            "maxLength": 500
          },
          "fields": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "结构化纠错的待采纳字段；为空时采纳全部字段"
          }
        }
      },
...
      "SuccessData": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "example": {
          "success": true
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "example": {
          "code": 10000,
          "message": "invalid request"
        }
      },
...
      "SpotCorrectionEnvelope": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "ok"
          },
          "data": {
            "$ref": "#/components/schemas/SpotCorrectionItem"
          }
        }
      },
      "ListSpotCorrectionsEnvelope": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "ok"
          },
          "data": {
            "$ref": "#/components/schemas/ListSpotCorrectionsResponse"
          }
        }
      },
      "ListMediaAssetsEnvelope": {
        "type": "object",
        "properties": {
          "code": {"type": "integer", "example": 0},
          "message": {"type": "string", "example": "ok"},
          "data": {"$ref": "#/components/schemas/ListMediaAssetsResponse"}
        }
      },
      "ListPublicSpotApplicationsEnvelope": {
        "type": "object",
        "properties": {
          "code": {"type": "integer", "example": 0},
          "message": {"type": "string", "example": "ok"},
          "data": {"$ref": "#/components/schemas/ListPublicSpotApplicationsResponse"}
        }
      },
      "PublicSpotApplicationEnvelope": {
        "type": "object",
        "properties": {
          "code": {"type": "integer", "example": 0},
          "message": {"type": "string", "example": "ok"},
          "data": {"$ref": "#/components/schemas/PublicSpotApplicationItem"}
        }
      },
...
      "SuccessEnvelope": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "ok"
          },
          "data": {
            "$ref": "#/components/schemas/SuccessData"
          }
        },
        "example": {
          "code": 0,
          "message": "ok",
          "data": {
            "success": true
          }
        }
      }
    }
  }
}